+++ news/apps.py
@@ -2,5 +2,9 @@ from django.apps import AppConfig
 
 
 class NewsConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'news'
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "news"
+
+    def ready(self):
+        # Import signal handlers
+        from . import signals  # noqa: F401